Changed camera culling to hide local player
I got annoyed by not being able to look down, so this is a possible fix
(though it looks weird in third person). I haven't tested it in
multiplayer, so it may be possible that it hides to other players too
depending on what layer they are spawned in. Hope to test soon